requirements.txt
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==2.0.12
colorama==0.4.6
flake8==5.0.4
flake8-docstrings==1.6.0
idna==3.10
iniconfig==2.0.0
mccabe==0.7.0
packaging==24.1
pluggy==1.5.0
py==1.11.0
pycodestyle==2.9.1
pydocstyle==6.3.0
pyflakes==2.5.0
pyTelegramBotAPI==4.14.1
pytest==7.1.3
pytest-timeout==2.1.0
python-dotenv==0.20.0
requests==2.26.0
snowballstemmer==2.2.0
tomli==2.0.2
types-requests==2.32.0.20241016
urllib3==2.2.3
